Rename nlp to lg and consolidate all modules.
After this change, there is only one module, the lg module,
which wraps all LG code. The concept of an nlp module is gone.
