+++ opencog/CMakeLists.txt
@@ -4,7 +4,7 @@ INCLUDE_DIRECTORIES(${CMAKE_BINARY_DIR})
 
 DECLARE_GUILE_CONFIG_TARGET(SCM_CONFIG "opencog lg-config" "OPENCOG_TEST")
 
-ADD_SUBDIRECTORY (nlp)
+ADD_SUBDIRECTORY (lg)
 
 WRITE_GUILE_CONFIG(${GUILE_BIN_DIR}/opencog/lg-config.scm SCM_CONFIG TRUE)
