+++ opencog/lg/cython/opencog/lg_types.pyx
@@ -2,9 +2,9 @@ from opencog.atomspace import get_refreshed_types
 from opencog.utilities import add_node, add_link
 
 cdef extern :
-    void nlp_types_init()
+    void lg_types_init()
 
-nlp_types_init()
-types = get_refreshed_types() 
+lg_types_init()
+types = get_refreshed_types()
 
-include "opencog/nlp/types/nlp_types.pyx"
+include "opencog/lg/types/lg_types.pyx"
