+++ opencog/lg/lg-conn/LGConnExpand.cc
@@ -22,7 +22,7 @@
 #include <opencog/atoms/atom_types/NameServer.h>
 #include <opencog/atoms/base/Node.h>
 #include <opencog/atomspace/AtomSpace.h>
-#include <opencog/nlp/types/atom_types.h>
+#include <opencog/lg/types/atom_types.h>
 #include "LGConnExpand.h"
 
 using namespace opencog;
@@ -171,3 +171,7 @@ ValuePtr LGConnExpand::execute(AtomSpace* as, bool silent)
 }
 
 DEFINE_LINK_FACTORY(LGConnExpand, LG_CONN_EXPAND)
+
+extern "C" {
+void opencog_nlp_lgconn_init(void) {}
+};
