+++ opencog/lg/lg-conn/LGConnExpand.h
@@ -23,7 +23,7 @@
 #define _OPENCOG_LG_CONN_EXPAND_H
 
 #include <opencog/atoms/core/FunctionLink.h>
-#include <opencog/nlp/types/atom_types.h>
+#include <opencog/lg/types/atom_types.h>
 
 namespace opencog
 {
