+++ opencog/lg/lg-dict/CMakeLists.txt
@@ -1,7 +1,7 @@
 
 INCLUDE_DIRECTORIES (
 	${LINK_GRAMMAR_INCLUDE_DIRS}	# for LinkGrammar dictionary
-	${CMAKE_BINARY_DIR}           # for the NLP atom types
+	${CMAKE_BINARY_DIR}           # for the LG atom types
 )
 
 ADD_LIBRARY (lg-dict-entry SHARED
@@ -16,10 +16,10 @@ ADD_LIBRARY (lg-dict SHARED
 	LGDictSCM.cc
 )
 
-ADD_DEPENDENCIES (lg-dict-entry nlp_atom_types)
+ADD_DEPENDENCIES (lg-dict-entry lg_atom_types)
 
 TARGET_LINK_LIBRARIES (lg-dict-entry
-	nlp-types
+	lg-types
 	${ATOMSPACE_smob_LIBRARY}
 	${LINK_GRAMMAR_LIBRARY}
 )
@@ -38,16 +38,11 @@ INSTALL (TARGETS lg-dict
 	DESTINATION "lib${LIB_DIR_SUFFIX}/opencog"
 )
 
-ADD_GUILE_MODULE (FILES
-	lg-dict.scm
-	MODULE_DESTINATION "${GUILE_SITE_DIR}/opencog/nlp/lg-dict"
-)
-
 ADD_GUILE_EXTENSION(SCM_CONFIG lg-dict "opencog-ext-path-lg-dict")
 
 INSTALL (FILES
 	LGDictEntry.h
 	LGDictNode.h
 	LGDictUtils.h
-	DESTINATION "include/opencog/nlp/lg-dict"
+	DESTINATION "include/opencog/lg/lg-dict"
 )
