+++ opencog/lg/lg-dict/LGDictEntry.cc
