+++ opencog/lg/lg-dict/LGDictReader.cc
@@ -25,7 +25,7 @@
 
 #include <opencog/atoms/base/Link.h>
 #include <opencog/atoms/base/Node.h>
-#include <opencog/nlp/types/atom_types.h>
+#include <opencog/lg/types/atom_types.h>
 #include "LGDictReader.h"
 
 using namespace opencog;
