+++ opencog/lg/lg-dict/LGDictUtils.cc
@@ -24,7 +24,7 @@
 #include <opencog/atomspace/AtomSpace.h>
 #include <opencog/atoms/base/Link.h>
 #include <opencog/atoms/base/Node.h>
-#include <opencog/nlp/types/atom_types.h>
+#include <opencog/lg/types/atom_types.h>
 
 #include "LGDictUtils.h"
