+++ opencog/lg/lg-dict/LGDictUtils.h
