+++ opencog/lg/lg-parse/CMakeLists.txt
@@ -2,18 +2,18 @@
 INCLUDE_DIRECTORIES (
 	${LINK_GRAMMAR_INCLUDE_DIRS}	# for LinkGrammar dictionary
 	${UUID_INCLUDE_DIRS}
-	${CMAKE_BINARY_DIR}           # for the NLP atom types
+	${CMAKE_BINARY_DIR}           # for the LG atom types
 )
 
 ADD_LIBRARY (lg-parse SHARED
 	LGParseLink.cc
 )
 
-ADD_DEPENDENCIES (lg-parse nlp_atom_types)
+ADD_DEPENDENCIES (lg-parse lg_atom_types)
 
 TARGET_LINK_LIBRARIES (lg-parse
 	lg-dict-entry
-	nlp-types
+	lg-types
 	${ATOMSPACE_STORAGE_LIBRARIES}
 	${ATOMSPACE_LIBRARIES}
 	${LINK_GRAMMAR_LIBRARY}
@@ -23,11 +23,6 @@ TARGET_LINK_LIBRARIES (lg-parse
 INSTALL (TARGETS lg-parse
 	EXPORT LGAtomeseTargets
 	DESTINATION "lib${LIB_DIR_SUFFIX}/opencog"
-)
-
-ADD_GUILE_MODULE (FILES
-	lg-parse.scm
-	MODULE_DESTINATION "${GUILE_SITE_DIR}/opencog/nlp/lg-parse"
 )
 
 ADD_GUILE_EXTENSION(SCM_CONFIG lg-parse "opencog-ext-path-lg-parse")
