+++ opencog/lg/lg-parse/LGParseLink.cc
