+++ opencog/lg/scm/CMakeLists.txt
@@ -1,0 +1,4 @@
+ADD_GUILE_MODULE (FILES
+	lg.scm
+	MODULE_DESTINATION "${GUILE_SITE_DIR}/opencog/lg"
+)
