+++ opencog/lg/scm/attic/README.md
