+++ opencog/lg/scm/attic/lg-dict-utils.scm
