+++ opencog/lg/scm/attic/nlp-utils.scm
