+++ opencog/lg/types/NLPTypes.cc
@@ -1,0 +1,14 @@
+/**
+ * NLPTypes.cc
+ *
+ * Atom Types used during Link Grammar processing.
+ *
+ * Copyright (c) 2009, 2014 Linas Vepstas <linasvepstas@gmail.com>
+ */
+
+#include "opencog/lg/types/atom_types.definitions"
+
+#define INHERITANCE_FILE "opencog/lg/types/atom_types.inheritance"
+#define INITNAME lg_types_init
+
+#include <opencog/atoms/atom_types/atom_types.cc>
