+++ opencog/lg/types/atom_types.script
