--- opencog/nlp/lg-parse/lg-parse.scm
@@ -1,10 +1,0 @@
-;
-; lg-parse.scm
-;
-; Link Grammar Parsing API
-;
-(define-module (opencog nlp lg-parse))
-
-(use-modules (opencog) (opencog lg-config) (opencog nlp))
-
-(load-extension (string-append opencog-ext-path-lg-parse "liblg-parse") "opencog_nlp_lgparse_init")
