--- opencog/nlp/scm/CMakeLists.txt
@@ -1,4 +1,0 @@
-ADD_GUILE_MODULE (FILES
-	nlp.scm
-	MODULE_DESTINATION "${GUILE_SITE_DIR}/opencog/nlp"
-)
