--- opencog/nlp/types/NLPTypes.cc
@@ -1,14 +1,0 @@
-/**
- * NLPTypes.cc
- *
- * Atom Types used during NLP processing.
- *
- * Copyright (c) 2009, 2014 Linas Vepstas <linasvepstas@gmail.com>
- */
-
-#include "opencog/nlp/types/atom_types.definitions"
-
-#define INHERITANCE_FILE "opencog/nlp/types/atom_types.inheritance"
-#define INITNAME nlp_types_init
-
-#include <opencog/atoms/atom_types/atom_types.cc>
